refactor: move literal-code to literal package
